backend/requirements.txt
# =============================================================================
# HACT CTMS — Python Dependencies
# =============================================================================

# --- Core Framework ---
Django==5.1.5
djangorestframework==3.15.2
django-filter==24.3

# --- Database ---
psycopg2-binary==2.9.10

# --- Authentication (Keycloak OIDC) ---
mozilla-django-oidc==4.0.1

# --- Async Task Queue ---
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# --- API Documentation ---
drf-spectacular==0.28.0

# --- CORS ---
django-cors-headers==4.6.0

# --- Production Server ---
gunicorn==23.0.0

# --- Utilities ---
python-dotenv==1.0.1
django-environ==0.11.2

# --- Health Checks ---
django-health-check==3.18.3
